Changed some more packets for making them work on both sides. Fixed windown building  process after making it work under  FreeBSD. Bumped  DevIL to 1.8.0 ( latest ) for both client & server.

+++ Lead-Client-Source/UserInterface/PythonBackground.cpp
@@ -682,13 +682,13 @@ void CPythonBackground::SetBackgroundDirLight()
 	STATEMANAGER.SetLight(0, &mc_pcurEnvironmentData->DirLights[ENV_DIRLIGHT_BACKGROUND]);
 }
 
-void CPythonBackground::GlobalPositionToLocalPosition(LONG& rGlobalX, LONG& rGlobalY)
+void CPythonBackground::GlobalPositionToLocalPosition(int32_t& rGlobalX, int32_t& rGlobalY)
 {
 	rGlobalX-=m_dwBaseX;
 	rGlobalY-=m_dwBaseY;
 }
 
-void CPythonBackground::LocalPositionToGlobalPosition(LONG& rLocalX, LONG& rLocalY)
+void CPythonBackground::LocalPositionToGlobalPosition(int32_t& rLocalX, int32_t& rLocalY)
 {
 	rLocalX+=m_dwBaseX;
 	rLocalY+=m_dwBaseY;
