+++ Lead-Client-Source/UserInterface/PythonBackground.h
@@ -61,8 +61,8 @@ public:
 	void Destroy();
 	void Create();
 
-	void GlobalPositionToLocalPosition(LONG& rGlobalX, LONG& rGlobalY);
-	void LocalPositionToGlobalPosition(LONG& rLocalX, LONG& rLocalY);
+	void GlobalPositionToLocalPosition(int32_t& rGlobalX, int32_t& rGlobalY);
+	void LocalPositionToGlobalPosition(int32_t& rLocalX, int32_t& rLocalY);
 
 	void EnableTerrainOnlyForHeight();
 	bool SetSplatLimit(int iSplatNum);
